+++ packages/mongo-nav/src/helpers/index.ts
@@ -1,4 +1,14 @@
 import OrgNavLink from './OrgNavLink';
 import InteractionRingWrapper from './InteractionRingWrapper';
 import OnPremUserMenu from './OnPremUserMenu';
-export { OrgNavLink, InteractionRingWrapper, OnPremUserMenu };
+import ProjectStatusBadge from './ProjectStatusBadge';
+import { RealmIcon, AtlasIcon, ChartsIcon } from './Icons';
+export {
+  OrgNavLink,
+  InteractionRingWrapper,
+  OnPremUserMenu,
+  ProjectStatusBadge,
+  RealmIcon,
+  AtlasIcon,
+  ChartsIcon,
+};
